Added implementation of dsVolatileMaterialValues.

+++ Core/include/DeepSea/Core/Thread/ConditionVariable.h
@@ -39,13 +39,13 @@ extern "C"
  * @brief Gets the size of dsConditionVariable.
  * @return The size of dsConditionVariable.
  */
-DS_CORE_EXPORT unsigned int dsConditionVariable_sizeof(void);
+DS_CORE_EXPORT size_t dsConditionVariable_sizeof(void);
 
 /**
  * @brief Gets the full allocated size of dsConditionVariable.
  * @return The full allocated size of dsConditionVariable.
  */
-DS_CORE_EXPORT unsigned int dsConditionVariable_fullAllocSize(void);
+DS_CORE_EXPORT size_t dsConditionVariable_fullAllocSize(void);
 
 /**
  * @brief Creates a condition variable.
